+++ src/main/java/org/sifenboot/security/certificado/repository/CertificadoRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 public interface CertificadoRepository extends JpaRepository<Certificado, Long> {
 
     // Busca por la PK (id) de la entidad Emisor
-    Optional<Certificado> findByEmisorId(Long emisorId);
+    Optional<Certificado> findByEmisorId(Integer emisorId);
 
     // Corregido: Spring navegará Certificado -> Emisor -> codEmisor
     Optional<Certificado> findByEmisorCodEmisor(String codEmisor);
